code

+++ app/Http/Controllers/Admin/BlogController.php
@@ -5,9 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\Blog;
 use Illuminate\Http\Request;
-use Illuminate\Support\Str;
-use Storage;
-use Image;
+use File;
 use Carbon\Carbon;
 use Auth;
 
@@ -47,33 +45,22 @@ class BlogController extends Controller
         //      'blog_title'      => 'required',
         //     ' blog_des'        => 'required',
         // ]);
-
-            $image = $request->file('blog_image');
-            $slug  = Str::slug($request->blog_title);
-         
-          if (isset($image))
-        {
-           //  make unique name for image
-            $currentDate = Carbon::now()->toDateString();
-            $imagename = $slug.'-'.$currentDate.'-'.uniqid().'.'.$image->getClientOriginalExtension();
-           //    check blog dir is exists
-            if (!Storage::disk('public')->exists('blog'))
-            {
-                Storage::disk('public')->makeDirectory('blog');
-            }
-         //  resize image for blog and upload
-        
-            Image::make($image)->resize(626, 417)->save(storage_path('app/public/blog').'/'.$imagename);
-
-        }
-
-
+     
         $blog = new Blog();
-        $blog->blog_image       = $imagename;
         $blog->blog_title       = $request->blog_title;
         $blog->added_by         = Auth::id();
         $blog->blog_des         = $request->blog_des;
-        $blog->save();
+
+         if ($request->hasFile('blog_image')) {
+                $file = $request->file('blog_image');
+                $extension = $file->getClientOriginalExtension();
+                $filename = time().".".$extension;
+                $file->move('uploads/blog/',$filename);
+                $blog->blog_image = $filename;
+
+            }
+
+            $blog->save();
 
         notify()->success("Blog Added", "Success");
         return redirect()->route('admin.blog.index');
@@ -116,36 +103,22 @@ class BlogController extends Controller
         //     ' blog_des'        => 'required',
         // ]);
 
-            $image = $request->file('blog_image');
-            $slug  = Str::slug($request->blog_title);
-         
-          if (isset($image))
-        {
-           //  make unique name for image
-            $currentDate = Carbon::now()->toDateString();
-            $imagename = $slug.'-'.$currentDate.'-'.uniqid().'.'.$image->getClientOriginalExtension();
-           //    check blog dir is exists
-            if (!Storage::disk('public')->exists('blog'))
-            {
-                Storage::disk('public')->makeDirectory('blog');
-            }
-
-            if (Storage::disk('public')->exists('blog/'.$blog->blog_image)) {
-            Storage::disk('public')->delete('blog/'.$blog->blog_image);
-            }
-
-
-         //  resize image for blog and upload
-        
-            Image::make($image)->resize(626, 417)->save(storage_path('app/public/blog').'/'.$imagename);
-
-        }else{
-            $imagename = $blog->blog_image;
-        }
-
         $blog->blog_image       = $imagename;
         $blog->blog_title       = $request->blog_title;
         $blog->blog_des         = $request->blog_des;
+          if ($request->hasFile('blog_image')) {
+
+             if (File::exists($image_path)) {
+                    File::delete($image_path);
+                }
+
+                $file = $request->file('blog_image');
+                $extension = $file->getClientOriginalExtension();
+                $filename = time().".".$extension;
+                $file->move('uploads/blog/',$filename);
+                $blog->blog_image = $filename;
+
+            }
         $blog->save();
 
         notify()->success("Blog Added", "Success");
@@ -160,9 +133,9 @@ class BlogController extends Controller
      */
     public function destroy(Blog $blog)
     {
-        if (Storage::disk('public')->exists('blog/'.$blog->blog_image)) {
-            Storage::disk('public')->delete('blog/'.$blog->blog_image);
-            }
+       if (File::exists($image_path)) {
+                    File::delete($image_path);
+                }
             
         $blog->delete();
         notify()->success('Blog Deleted', "Success");
